requirements.txt
requests
git+https://gitee.com/ikaros-521/WenxinWorkshop-Python-SDK
git+https://gitee.com/ikaros-521/blivedm
git+https://gitee.com/ikaros-521/sparkdesk-api
openai
pygame
edge-tts==6.1.6
langid
aiohttp==3.9.1
asyncio
slack_sdk==3.21.3
profanity==1.1
elevenlabs==0.2.8
PyPDF2
protobuf_inspector
websocket-client
websocket-client-py3
tiktoken
pyahocorasick
webuiapi
pyvirtualcam
numpy
faiss-cpu
sentence_transformers
pydub
pydantic==1.10.13
send2trash
keyboard
PyAudio
SpeechRecognition
baidu-aip
chardet
pypinyin
gradio_client
schedule
bilibili-api-python
langchain==0.0.142
playwright
protobuf==3.20.3
certifi
charset-normalizer
greenlet
idna
lxml
pyee
typing_extensions
urllib3
pyhandytools
curl_cffi
zhipuai
bardapi
revTongYi
pyautogui
nicegui
pytchat
PySocks
pure-protobuf==3.0.0a5
flask
flask_cors
xingchen
flask_socketio
emoji
google-generativeai
colorlog
faster_whisper
httpx==0.25.2
Pillow==10.1.0
pygtrans
jieba
gradio==4.16.0
TikTokLive
azure-cognitiveservices-speech
pyjwt
dashscope
pygetwindow
opencv-python
g4f
starlette==0.27.0
pyserial

# dy2
betterproto==2.0.0b6

loguru